// src/credit/dto/create-credit.dto.ts
export class CreateCreditDto {
    old_price: number;
    new_price: number;
    customer_id: number;
    change_price: number;
    note: string;
    reason: string;
    receiver_id: number;
    store_id: number;
    store_name: string;
    created_by: number;
    created_by_name: string;
    order_code: number;
    package_code: string;
}
  